NM: move votes into vote scraper (WIP).
