+++ openstates/nm/bills.py
@@ -3,7 +3,6 @@ import re
 import csv
 import zipfile
 import subprocess
-import operator
 from datetime import datetime
 from bisect import bisect_left
 
@@ -11,16 +10,7 @@ import lxml.html
 import lxml.etree
 import scrapelib
 
-from pupa.scrape import Scraper, Bill, VoteEvent
-from pupa.utils.generic import convert_pdf
-
-# Senate vote header
-s_vote_header = re.compile(r'(YES)|(NO)|(ABS)|(EXC)|(REC)')
-# House vote header
-h_vote_header = re.compile(r'(YEA(?!S))|(NAY(?!S))|(EXCUSED(?!:))|(ABSENT(?!:))')
-
-# Date regex for senate and house parser
-date_regex = re.compile(r'([0-1][0-9]\/[0-3][0-9]\/\d+)')
+from pupa.scrape import Scraper, Bill
 
 
 def convert_sv_char(c):
@@ -203,8 +193,6 @@ class NMBillScraper(Scraper):
         self.scrape_documents(session, 'bills', chamber, bills)
         self.scrape_documents(session, 'resolutions', chamber, bills)
         self.scrape_documents(session, 'memorials', chamber, bills)
-        self.scrape_documents(session, 'votes', chamber, bills,
-                              chamber_name='')
         self.check_other_documents(session, chamber, bills)
         # self.dedupe_docs(bills)
 
@@ -242,12 +230,12 @@ class NMBillScraper(Scraper):
                                 'document for unknown bill {}'.format(fname))
                         else:
                             if doc_type == 'Final Version':
-                                bill.add_version(
+                                bill.add_version_link(
                                     'Final Version', url + fname,
-                                    mimetype=mimetype)
+                                    media_type=mimetype)
                             else:
                                 bill.add_document_link(doc_type, url + fname,
-                                                       mimetype=mimetype)
+                                                       media_type=mimetype)
 
         check_docs(firs_url, 'Fiscal Impact Report')
         check_docs(lesc_url, 'LESC Analysis')
@@ -477,31 +465,12 @@ class NMBillScraper(Scraper):
                 committee_name = suffix[:-1]
                 bill.add_version_link('%s substitute' % committee_name,
                                       doc_path + fname, media_type=media_type)
-            # votes
-            elif 'SVOTE' in suffix:
-                sv_text = self.scrape_vote(doc_path + fname)
-                if not sv_text:
-                    continue
-
-                vote = self.parse_senate_vote(sv_text, doc_path + fname, bill)
-                if not vote:
-                    self.warning(
-                        'Bad parse on the senate vote for {}'.format(bill_id))
-
-            elif 'HVOTE' in suffix:
-                hv_text = self.scrape_vote(doc_path + fname)
-                if not hv_text:
-                    continue
-                vote = self.parse_house_vote(hv_text, doc_path + fname, bill)
-                if not vote:
-                    self.warning(
-                        'Bad parse on the house vote for {}'.format(bill_id))
 
             # committee reports
             elif re.match(r'\w{2,4}\d', suffix):
                 committee_name = re.match(r'[A-Z]+', suffix).group()
                 bill.add_document_link('%s committee report' % committee_name,
-                                       doc_path + fname, mimetype=media_type)
+                                       doc_path + fname, media_type=media_type)
 
             # ignore list, mostly typos reuploaded w/ proper name
             elif suffix in ('HEC', 'HOVTE', 'GUI'):
@@ -512,258 +481,6 @@ class NMBillScraper(Scraper):
                 # duplicates
                 self.warning('unknown document suffix {} ({})'.format(suffix,
                                                                       fname))
-
-    def scrape_vote(self, filelocation, local=False):
-        """Retrieves or uses local copy of vote pdf and converts into XML."""
-        if not local:
-            try:
-                filename, response = self.urlretrieve(url=filelocation)
-                vote_text = convert_pdf(filename, type='xml')
-                os.remove(filename)
-            except scrapelib.HTTPError:
-                self.warning('Request failed: {}'.format(filelocation))
-                return
-        else:
-            vote_text = convert_pdf(filelocation, type='xml')
-            os.remove(filelocation)
-        return vote_text
-
-    def parse_house_vote(self, hv_text, url, bill):
-        """Sets any overrides and creates the vote instance"""
-        overrides = {'ONEILL': "O'NEILL"}
-        # Add new columns as they appear to be safe
-        vote_record, row_headers, sane_row = self.parse_visual_grid(
-            hv_text,
-            overrides,
-            h_vote_header,
-            'CERTIFIED CORRECT',
-            'YEAS')
-        vote = self.build_vote(bill, url, vote_record, 'lower',
-                               'house passage')
-
-        self.validate_house_vote(row_headers, sane_row, vote_record)
-        return vote
-
-    def parse_senate_vote(self, sv_text, url, bill):
-        """Sets any overrides and creates the vote instance"""
-        overrides = {'ONEILL': "O'NEILL"}
-        # Add new columns as they appear to be safe
-        vote_record, row_headers, sane_row = self.parse_visual_grid(
-            sv_text,
-            overrides,
-            s_vote_header,
-            'TOTAL',
-            'TOTAL')
-        vote = self.build_vote(bill, url, vote_record, 'upper',
-                               'senate passage')
-
-        self.validate_senate_vote(row_headers, sane_row, vote_record)
-        return vote
-
-    @staticmethod
-    def build_vote(bill, url, vote_record, chamber, motion_text):
-        passed = len(vote_record['yes']) > len(vote_record['no'])
-        vote = VoteEvent(result=passed,
-                         chamber=chamber,
-                         start_date=vote_record['date'],
-                         motion_text=motion_text,
-                         classification='passage',
-                         legislative_session=bill.legislative_session,
-                         bill=bill)
-        vote.set_count('yes', len(vote_record['yes']))
-        vote.set_count('no', len(vote_record['no']))
-        vote.set_count('excused', len(vote_record['excused']))
-        vote.set_count('absent', len(vote_record['absent']))
-        vote.set_count('other', len(vote_record['other']))
-        for vote_type in ['yes', 'no', 'excused', 'absent', 'other']:
-            for voter in vote_record[vote_type]:
-                vote.vote(vote_type, voter)
-
-        vote.add_source(url)
-        return vote
-
-    @staticmethod
-    def validate_house_vote(row_headers, sane_row, vote_record):
-        # Sanity checks on vote data, checks that the calculated total and
-        # listed totals match
-        sane = {'yea': 0, 'nay': 0, 'excused': 0, 'absent': 0, 'other': 0}
-        # Make sure the header row and sanity row are in order
-        sorted_row_header = sorted(row_headers.items(),
-                                   key=operator.itemgetter(0))
-        start_count = -1
-        for cell in sane_row:
-            cell_value = cell[0].split()[-1].strip()
-            if 'YEAS' in cell[0] or start_count >= 0:
-                start_count += 1
-                sane_vote = sorted_row_header[start_count][1]
-                if 'Y' == sane_vote[0]:
-                    sane['yea'] = int(cell_value)
-                elif 'N' == sane_vote[0]:
-                    sane['nay'] = int(cell_value)
-                elif 'E' == sane_vote[0]:
-                    sane['excused'] = int(cell_value)
-                elif 'A' == sane_vote[0]:
-                    sane['absent'] = int(cell_value)
-                else:
-                    sane['other'] += int(cell_value)
-        # Make sure the parsed vote totals match up with counts in the total
-        # field
-        if sane['yea'] != len(vote_record['yes']) or \
-            sane['nay'] != len(vote_record['no']) or \
-            sane['excused'] != len(vote_record['excused']) or \
-            sane['absent'] != len(vote_record['absent']) or \
-                sane['other'] != len(vote_record['other']):
-            raise ValueError('Votes were not parsed correctly')
-        # Make sure the date is a date
-        if not isinstance(vote_record['date'], datetime):
-            raise ValueError('Date was not parsed correctly')
-        # End Sanity Check
-
-    @staticmethod
-    def validate_senate_vote(row_headers, sane_row, vote_record):
-        # Sanity checks on vote data, checks that the calculated total and
-        # listed totals match
-        sane = {'yea': 0, 'nay': 0, 'excused': 0, 'absent': 0, 'other': 0}
-        # Make sure the header row and sanity row are in order
-        sorted_row_header = sorted(row_headers.items(),
-                                   key=operator.itemgetter(0))
-        start_count = -1
-        for cell in sane_row:
-            cell_value = cell[0]
-            if start_count >= 0:
-                sane_vote = sorted_row_header[start_count][1]
-                if 'Y' == sane_vote[0]:
-                    sane['yea'] = int(cell_value)
-                elif 'N' == sane_vote[0]:
-                    sane['nay'] = int(cell_value)
-                elif 'E' == sane_vote[0]:
-                    sane['excused'] = int(cell_value)
-                elif 'A' == sane_vote[0]:
-                    sane['absent'] = int(cell_value)
-                else:
-                    sane['other'] += int(cell_value)
-                start_count += 1
-            elif 'TOTAL' in cell_value:
-                start_count = 0
-        # Make sure the parsed vote totals match up with counts in the
-        # total field
-        if sane['yea'] != len(vote_record['yes']) or \
-            sane['nay'] != len(vote_record['no']) or \
-            sane['excused'] != len(vote_record['excused']) or \
-            sane['absent'] != len(vote_record['absent']) or \
-                sane['other'] != len(vote_record['other']):
-            raise ValueError('Votes were not parsed correctly')
-        # Make sure the date is a date
-        if not isinstance(vote_record['date'], datetime):
-            raise ValueError('Date was not parsed correctly')
-
-    def parse_visual_grid(self, v_text, overrides, vote_header,
-                          table_stop, sane_iden):
-        """
-        Takes a (badly)formatted pdf and converts the vote grid into an X,Y
-        grid to match votes
-        """
-        vote_record = {
-            'date': None,
-            'yes': [],
-            'no': [],
-            'excused': [],
-            'absent': [],
-            'other': []
-        }
-        row_heads = {}
-        column_map = {}
-        rows = {}
-        t_begin = 0
-        t_stop = 0
-        sane_row = 0
-        # Take the mixed up text tag cells and separate header/special and
-        # non-header cells.
-        # Metadata hints that this doc is done by hand, tags appear in
-        # chrono order not visual
-        for tag in lxml.etree.XML(v_text).xpath('//text/b') + lxml.etree.XML(
-                v_text).xpath('//text'):
-            if tag.text is None:
-                continue
-            row_value = tag.text.strip()
-            if 'top' not in tag.keys():
-                tag = tag.getparent()
-            top = int(tag.attrib['top'])
-            # name overrides
-            if row_value in overrides:
-                row_value = overrides[row_value]
-            elif 'LT. GOV' in row_value:
-                # Special case for the senate, inconsistencies make overrides
-                #  not an option
-                row_value = 'LT. GOVERNOR'
-            elif table_stop in row_value:
-                # Set the data table end point
-                t_stop = top
-
-            if sane_iden in row_value:
-                # Vote sanity row can be the same as the tableStop
-                sane_row = top
-            if date_regex.search(row_value):
-                # Date formats change depending on what document is being used
-                if len(row_value) == 8:
-                    vote_record['date'] = datetime.strptime(
-                        date_regex.search(row_value).group(), '%m/%d/%y')
-                else:
-                    vote_record['date'] = datetime.strptime(
-                        date_regex.search(row_value).group(), '%m/%d/%Y')
-            elif vote_header.match(row_value):
-                row_heads[int(tag.attrib['left']) + int(
-                    tag.attrib['width'])] = row_value
-                # Set the header begin sanity value
-                if t_begin == 0:
-                    t_begin = top
-            else:
-                # Create dictionary of row params and x,y
-                # location- y:{value, x, x(offset)}
-                if top in rows:
-                    rows[top].append((row_value, int(tag.attrib['left']),
-                                      int(tag.attrib['width'])))
-                else:
-                    rows[top] = [(row_value, int(tag.attrib['left']),
-                                  int(tag.attrib['width']))]
-
-        # Mark the votes in the datagrid
-        for row_x, cells in rows.items():
-            if t_begin < row_x <= t_stop:
-                # Resort the row cells to go left to right, due to possile
-                # table pane switching
-                cells.sort(key=operator.itemgetter(1))
-                # Each vote grid is made up of split tables with two active
-                # columns
-                for x in range(0, len(cells), 2):
-                    if table_stop in cells[x][0]:
-                        break
-                    if x + 1 >= len(cells):
-                        self.warning('No vote found for {}'.format(cells[x]))
-                        continue
-                    if cells[x+1][1] not in column_map:
-                        # Called one time for each column heading
-                        # Map the data grid column to the header columns
-                        column_map[cells[x+1][1]] = \
-                            match_header(list(row_heads.keys()),
-                                         cells[x + 1][1] + cells[x + 1][2])
-                    vote_cast = row_heads[column_map[cells[x+1][1]]]
-
-                    # Fix some odd encoding issues
-                    name = correct_name(''.join(convert_sv_char(c) for c in
-                                                cells[x][0]))
-                    if 'Y' == vote_cast[0]:
-                        vote_record['yes'].append(name)
-                    elif 'N' == vote_cast[0]:
-                        vote_record['no'].append(name)
-                    elif 'E' == vote_cast[0]:
-                        vote_record['excused'].append(name)
-                    elif 'A' == vote_cast[0]:
-                        vote_record['absent'].append(name)
-                    else:
-                        vote_record['other'].append(name)
-
-        return vote_record, row_heads, rows[sane_row]
 
     # def dedupe_docs(self, bills):
     #     for bill_id, bill in bills.items():
